+++ src/simulate/simulate_fusion_at_array.cpp
@@ -43,8 +43,8 @@ namespace das {
             DAS_PROFILE_NODE \
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = uint32_t(r.subexpr->evalInt(context)); \
-            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %u", rr, pl->size); \
-            return *((CTYPE *)(pl->data + rr*stride + offset)); \
+            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
+            return *((CTYPE *)(pl->data + uint64_t(rr)*uint64_t(stride) + offset)); \
         } \
         DAS_NODE(TYPE,CTYPE); \
     };
@@ -55,8 +55,8 @@ namespace das {
             DAS_PROFILE_NODE \
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = *((uint32_t *)r.compute##COMPUTER(context)); \
-            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %u", rr, pl->size); \
-            return *((CTYPE *)(pl->data + rr*stride + offset)); \
+            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
+            return *((CTYPE *)(pl->data + uint64_t(rr)*uint64_t(stride) + offset)); \
         } \
         DAS_NODE(TYPE,CTYPE); \
     };
@@ -84,9 +84,9 @@ namespace das {
             DAS_PROFILE_NODE \
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = uint32_t(r.subexpr->evalInt(context)); \
-            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %u", rr, pl->size); \
+            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
             vec4f __r; \
-            DAS_LDU_WORKHORSE(__r, pl->data + rr*stride + offset, CTYPE); \
+            DAS_LDU_WORKHORSE(__r, pl->data + uint64_t(rr)*uint64_t(stride) + offset, CTYPE); \
             return __r; \
         } \
     };
@@ -98,9 +98,9 @@ namespace das {
             DAS_PROFILE_NODE \
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = *((uint32_t *)r.compute##COMPUTER(context)); \
-            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %u", rr, pl->size); \
+            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
             vec4f __r; \
-            DAS_LDU_WORKHORSE(__r, pl->data + rr*stride + offset, CTYPE); \
+            DAS_LDU_WORKHORSE(__r, pl->data + uint64_t(rr)*uint64_t(stride) + offset, CTYPE); \
             return __r; \
         } \
     };
@@ -119,8 +119,8 @@ namespace das {
             DAS_PROFILE_NODE \
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = uint32_t(r.subexpr->evalInt(context)); \
-            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %u", rr, pl->size); \
-            return pl->data + rr*stride + offset; \
+            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
+            return pl->data + uint64_t(rr)*uint64_t(stride) + offset; \
         } \
         DAS_PTR_NODE; \
     };
@@ -132,8 +132,8 @@ namespace das {
             DAS_PROFILE_NODE \
             auto pl = (Array *) l.compute##COMPUTEL(context); \
             auto rr = *((uint32_t *)r.compute##COMPUTER(context)); \
-            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %u", rr, pl->size); \
-            return pl->data + rr*stride + offset; \
+            if ( rr >= pl->size ) context.throw_error_at(debugInfo,"array index out of range, %u of %llu", rr, (unsigned long long)pl->size); \
+            return pl->data + uint64_t(rr)*uint64_t(stride) + offset; \
         } \
         DAS_PTR_NODE; \
     };
